fix: burger btn backs to original position when menu closed
