+++ src/components/MobileMenu/MobileMenu.tsx
@@ -1,155 +1,52 @@
+/* eslint-disable jsx-a11y/no-static-element-interactions */
+/* eslint-disable jsx-a11y/click-events-have-key-events */
 import cl from 'classnames';
-import { useAtom } from 'jotai';
-import { FC, useEffect, useRef, useState } from 'react';
-import { NavLink, useLocation, useNavigate } from 'react-router-dom';
-import { useUserProfileQuery } from '../../api/queries';
-import { isAuthAtom } from '../../atoms/isAuthAtom';
-import {
-  ROUTE_COMPETENCIES,
-  ROUTE_CONTACT,
-  ROUTE_ED_TECH,
-  ROUTE_HOME,
-  ROUTE_PRODUCTS,
-  ROUTE_PROFILE,
-} from '../../constants/routesConstants';
-import { HamburgerBtn } from '../Header/HamburgerBtn/HamburgerBtn';
-import { ArrowIcon } from '../ui/icons';
+import { FC, useEffect, useState } from 'react';
+import { useLocation } from 'react-router';
+import { MobileNav } from './MobileNav/MobileNav';
 import s from './MobileMenu.module.scss';
-import { CompetenciesMobSubMenu } from './MobileSubMenu/CompetenciesMobSubMenu';
-import { ProductsMobSubMenu } from './MobileSubMenu/ProductsMobSubMenu';
 
-type Props = {
-  isOpen: boolean;
-  onBurgerClick: () => void;
-};
+type Props = {};
 
-export const MobileMenu: FC<Props> = ({ isOpen, onBurgerClick }) => {
-  const [isCompetenciesVisible, setCompetenciesVisible] = useState(false);
-  const [isProductsVisible, setProductsVisible] = useState(false);
-  const [scrollPosition, setScrollPosition] = useState(0);
+const MobileMenu: FC<Props> = () => {
+  const [isNavMobileOpen, setIsNavMobileOpen] = useState(true);
   const location = useLocation();
-  const navigate = useNavigate();
-  const [isAuth] = useAtom(isAuthAtom);
-  const mobileMenuRef = useRef<HTMLDivElement>(null);
-
-  const { data: userProfile } = useUserProfileQuery();
-
-  const isCompetenciesPage = location.pathname.includes(ROUTE_COMPETENCIES);
-  const isProductsPage = location.pathname.includes(ROUTE_PRODUCTS);
 
   useEffect(() => {
-    if (isOpen) {
+    // Close the menu when the location changes
+    setIsNavMobileOpen(false);
+  }, [location]);
+
+  // Add class to body to disable scrollbar when the popup is open
+  useEffect(() => {
+    if (isNavMobileOpen) {
       document.body.classList.add(s.bodyNoScroll);
     } else {
       document.body.classList.remove(s.bodyNoScroll);
     }
-  }, [isOpen]);
+  }, [isNavMobileOpen]);
 
-  // Add an event listener to the MobileMenu component to track the scroll position
-  useEffect(() => {
-    const handleScroll = () => {
-      if (mobileMenuRef.current) {
-        setScrollPosition(mobileMenuRef.current.scrollTop);
-      }
-    };
-
-    if (isOpen && mobileMenuRef.current) {
-      mobileMenuRef.current.addEventListener('scroll', handleScroll);
-    }
-
-    return () => {
-      if (mobileMenuRef.current) {
-        mobileMenuRef.current.removeEventListener('scroll', handleScroll);
-      }
-    };
-  }, [isOpen]);
-
-  const handleCompetenciesBtnClick = () => {
-    // On mobile there is no /competencies page.
-    if (window.innerWidth > 768) {
-      navigate(ROUTE_COMPETENCIES);
-    } else {
-      setCompetenciesVisible(!isCompetenciesVisible);
-    }
+  const handleBurgerBtnClick = () => {
+    setIsNavMobileOpen(!isNavMobileOpen);
   };
 
-  const handleProductsBtnClick = () => {
-    // On mobile there is no /competencies page.
-    if (window.innerWidth <= 768) {
-      setProductsVisible(!isProductsVisible);
+  const onOverlayTap = (e: React.MouseEvent) => {
+    if (e.currentTarget === e.target) {
+      setIsNavMobileOpen(false);
     }
   };
 
   return (
-    <>
-      <HamburgerBtn isChecked={isOpen} onClick={onBurgerClick} positionY={-scrollPosition} />
-
-      <nav
-        aria-label="Основное меню"
-        className={cl(s.nav, {
-          [s.navMobileVisible]: isOpen,
+    <div className={cl(s.mobMenuContainer, { [s.fullScreen]: isNavMobileOpen })}>
+      <div
+        className={cl(s.mobileOverlay, {
+          [s.mobileOverlayOpen]: isNavMobileOpen,
         })}
-        ref={mobileMenuRef}>
-        <ul className={cl(s.list)}>
-          <li className={cl(s.listItem, s.itemWithHover)}>
-            <NavLink to={ROUTE_HOME} className={({ isActive }) => (isActive ? s.linkActive : '')}>
-              Главная
-            </NavLink>
-          </li>
-
-          <li className={cl(s.listItem, s.listItemSubMenu)}>
-            <button
-              type="button"
-              onClick={handleCompetenciesBtnClick}
-              className={cl(s.btnSubmenu, {
-                [s.linkActive]: isCompetenciesPage,
-              })}>
-              Компетенции
-              <ArrowIcon className={cl(s.arrow, { [s.arrow_rotate]: isCompetenciesVisible })} />
-            </button>
-            <CompetenciesMobSubMenu isVisible={isCompetenciesVisible} />
-          </li>
-
-          <li className={cl(s.listItem, s.itemWithHover)}>
-            <NavLink
-              to={ROUTE_ED_TECH}
-              className={({ isActive }) => (isActive ? s.linkActive : '')}>
-              Лаборатория
-            </NavLink>
-          </li>
-
-          <li className={cl(s.listItem, s.listItemSubMenu)}>
-            <button
-              type="button"
-              onClick={handleProductsBtnClick}
-              className={cl(s.btnSubmenu, {
-                [s.linkActive]: isProductsPage,
-              })}>
-              Продукты
-              <ArrowIcon className={cl(s.arrow, { [s.arrow_rotate]: isProductsVisible })} />
-            </button>
-            <ProductsMobSubMenu isVisible={isProductsVisible} />
-          </li>
-
-          <li className={cl(s.listItem, s.itemWithHover)}>
-            <NavLink
-              to={ROUTE_CONTACT}
-              className={({ isActive }) => (isActive ? s.linkActive : '')}>
-              Контакты
-            </NavLink>
-          </li>
-
-          <li className={cl(s.listItem, s.itemWithHover)}>
-            <NavLink
-              to={ROUTE_PROFILE}
-              className={({ isActive }) => (isActive ? s.linkActive : '')}>
-              {isAuth && userProfile
-                ? `${userProfile.last_name} ${userProfile.first_name[0]}.`
-                : 'Личный кабинет'}
-            </NavLink>
-          </li>
-        </ul>
-      </nav>
-    </>
+        onClick={onOverlayTap}
+      />
+      <MobileNav isOpen={isNavMobileOpen} onBurgerClick={handleBurgerBtnClick} />
+    </div>
   );
 };
+
+export default MobileMenu;
