--- src/components/MobileMenuPortal/MobileMenuPortal.module.scss
@@ -1,48 +1,0 @@
-.mobMenuContainer {
-  display: none;
-
-  @media screen and (width <= 768px) {
-    position: fixed;
-    z-index: 11;
-
-    display: block;
-
-    width: 100%;
-    min-height: 100vh;
-    max-height: 100vh;
-  }
-}
-
-.mobileOverlay {
-  top: 0;
-  left: 0;
-
-  display: flex;
-
-  box-sizing: border-box;
-  width: 100%;
-  min-height: 100vh;
-  max-height: 100vh;
-
-  visibility: hidden;
-  opacity: 0;
-  background-color: var(--shadow);
-
-  transition: opacity 0.3s ease;
-
-  &Open {
-    @media screen and (width <= 768px) {
-      visibility: visible;
-      opacity: 1;
-    }
-  }
-}
-
-.bodyNoScroll {
-  overflow: hidden;
-
-  &::-webkit-scrollbar {
-    // Hide scrollbar on the body
-    display: none;
-  }
-}
